feat: add classify comp for icon size
