+++ sites/geohub/src/components/maplibre/symbol/IconSize.svelte
@@ -1,45 +1,68 @@
 <script lang="ts">
-	import NumberInput from '$components/util/NumberInput.svelte';
-	import { MAPSTORE_CONTEXT_KEY, type MapStore } from '$stores';
+	import { page } from '$app/stores';
+	import VectorValueClassification from '$components/maplibre/vector/VectorValueClassification.svelte';
+	import type { SpriteImage, VectorTileMetadata } from '$lib/types';
+	import {
+		MAPSTORE_CONTEXT_KEY,
+		SPRITEIMAGE_CONTEXT_KEY,
+		type MapStore,
+		type SpriteImageStore
+	} from '$stores';
+	import chroma from 'chroma-js';
+	import { hexToCSSFilter } from 'hex-to-css-filter';
 	import type { LayerSpecification } from 'maplibre-gl';
 	import { getContext, onMount } from 'svelte';
 
 	const map: MapStore = getContext(MAPSTORE_CONTEXT_KEY);
+	const spriteImageList: SpriteImageStore = getContext(SPRITEIMAGE_CONTEXT_KEY);
 
 	export let layerId: string;
+	export let metadata: VectorTileMetadata;
+	export let defaultColor: string;
 
-	const propertyName = 'icon-size';
-	const style = $map
-		.getStyle()
-		.layers.filter((layer: LayerSpecification) => layer.id === layerId)[0];
-
+	let defaultLineWidth = $page.data.config.LineWidth;
 	let maxValue = 5;
 	let minValue = 0;
+	let propertyName = 'icon-size';
 	let stepValue = 0.25;
 
-	const getValue = () => {
-		if (style?.layout && style.layout[propertyName]) {
-			const value = style.layout[propertyName];
-			if (typeof value === 'object' && 'stops' in value) {
-				return 1;
-			}
-			return value;
-		} else {
-			return 1;
-		}
+	let icon: SpriteImage;
+	let cssIconFilter = '';
+
+	const setCssIconFilter = () => {
+		const rgba = chroma(defaultColor).rgba();
+		cssIconFilter = hexToCSSFilter(chroma([rgba[0], rgba[1], rgba[2]]).hex()).filter;
 	};
 
-	let value = getValue();
+	const getIconImageName = () => {
+		const propertyName = 'icon-image';
+		const style = $map
+			.getStyle()
+			.layers.filter((mapLayer: LayerSpecification) => mapLayer.id === layerId)[0];
+		return style.layout && style.layout[propertyName] ? style.layout[propertyName] : 'circle';
+	};
 
 	onMount(() => {
-		setValue();
+		handleDefaultColorChanged();
 	});
 
-	const setValue = () => {
-		map.setLayoutProperty(layerId, propertyName, value);
+	$: defaultColor, handleDefaultColorChanged();
+	const handleDefaultColorChanged = () => {
+		icon = $spriteImageList.find((icon) => icon.alt === getIconImageName());
+		setCssIconFilter();
 	};
 </script>
 
-<div data-testid="icon-size-input">
-	<NumberInput bind:value bind:minValue bind:maxValue bind:step={stepValue} on:change={setValue} />
-</div>
+{#if icon}
+	<VectorValueClassification
+		{layerId}
+		{metadata}
+		bind:defaultValue={defaultLineWidth}
+		{minValue}
+		{maxValue}
+		{stepValue}
+		{propertyName}
+		styleType="layout"
+		legendCssTemplate={`margin-left: auto; margin-right: auto; width: calc(1em * {value}); height: calc(1em * {value}); filter: ${cssIconFilter}; background-image: url("${icon.src}"); background-repeat: no-repeat; background-size: contain;`}
+	/>
+{/if}
