+++ sites/geohub/src/components/maplibre/symbol/VectorSymbol.svelte
@@ -2,9 +2,13 @@
 	import IconColor from '$components/maplibre/symbol/IconColor.svelte';
 	import IconImage from '$components/maplibre/symbol/IconImage.svelte';
 	import FieldControl from '$components/util/FieldControl.svelte';
-	import { handleEnterKey } from '$lib/helper';
+	import { getVectorDefaultColor, handleEnterKey } from '$lib/helper';
 	import type { VectorTileMetadata } from '$lib/types';
+	import { MAPSTORE_CONTEXT_KEY, type MapStore } from '$stores';
+	import { getContext, onMount } from 'svelte';
 	import IconSize from './IconSize.svelte';
+
+	const map: MapStore = getContext(MAPSTORE_CONTEXT_KEY);
 
 	export let layerId: string;
 	export let metadata: VectorTileMetadata;
@@ -25,6 +29,16 @@
 		}
 	];
 	let activeTab: string = tabs[0].id;
+
+	onMount(() => {
+		setDefaultColor();
+	});
+
+	$: activeTab, setDefaultColor();
+
+	const setDefaultColor = () => {
+		defaultColor = getVectorDefaultColor($map, layerId, 'icon-color');
+	};
 </script>
 
 <div class="tabs is-centered">
@@ -72,7 +86,7 @@
 	<FieldControl title="Icon size">
 		<div slot="help">Change icon color by using single color or selected property.</div>
 		<div slot="control">
-			<IconSize {layerId} />
+			<IconSize {layerId} {metadata} bind:defaultColor />
 		</div>
 	</FieldControl>
 </div>
